Use `AuthProvider` as single source of truth
Different wallets might have different authentication methods e.g. API
Keys or OAuth2. The `AuthProvider` handles the authentication logic,
e.g refreshing the access token for OAuth2 and always provides updated
authentication. In case a request still fails with 401 this is not
ambiguous, the error will be passed to the user.