common: Rename `Invoice` to `Bolt11Invoice`
This class is used for type-safety checks when passing around bolt11
invoices, and also to parse the actual bolt11 invoice to access things
like the amount easily. In the future we will implement a generic
"Invoice" class / model which will be Bitcoin invoices in different
formats (URI, BOLT11, LNURL) so we rename it now to avoid confusion.