common: Pass executor service to Barcode Scanner
MlKit Barcode Scanner has overloads for attaching the listeners
(`addOnSuccessListener` etc...) so far we did not pass an executor to
these functions which meant the barcode scanner used the main thread. It
is probably better to explicitly pass an executor there, so we can also
cancel/shutdown tasks on cleanup when the view is destroyed.