domain/data: Refactor `WalletRepository` methods
Refactor the `WalletRefactory` interface methods to take
`Invoice.Bolt11` as arguements instead of the `Bolt11Invoice` that is
contained in them. This also indicates that the `Bolt11Invoice` class is
probably not really usefull anymore and can be consolidated into the
`Invoice` type/class as its sole purpose was to "guard" the constructor
so that `encodedSafe` was always guaranteed to be valid eventhough its
just a raw string.