presentation: Use UI Layer models
Using Data Layer models or even DTO's is tempting because the UI Layer
is interested in mostly the same shape/data most of the time. But doing
this lead to really ugly code/architecture so now I'll start using UI
Layer models and the use cases must map between domain/data layer
models.

There might be a few domain/data model types left in the UI or the UI
models respecitvely but this will be cleaned up. Also the Strings in the
PaymentData and as the fee probe use case return are ugly but I will
implement a formatable type for this in the future.