presentation: Type `UiAmount`
Add `UiAmount` type/model for UI Layer. It is hardcoded to the
`UiAmount.Sat(..)` for now. Next we'll implement a use case to create
these amounts based on currency setting, pass a locale or rather number
formatter and also fetch exchange rate if a fiat amount needs to be
created.