domain/presentation: Refactor UI amount handling
The previous implementation worked but was quite awkward. This refactor
implements better types for display currencies and introduces a
`Formatter` interface to format different kinds of currencies.

The `GetExchangeRateFlowUseCase` now simply polls the repository but
maintains a cache of the already created flows so that new consumers
consume the same flows. The `GetFormatterFlowUseCase` seems like it
violates SRP but this is intentional, we want to update UI, or rather
the formatting of amounts when:
- locale changes
- currency changes
- rate changes

Putting the rate in there is kind of crazy but I thought it'd be cool.